requirements.txt
appium-python-client~=2.7.1
selenium~=4.5.0
behave~=1.2.6
pytest~=7.1.1